+++ change_log.md
@@ -16,3 +16,5 @@
 #### ver 0.0.7 Rescale u,v amplitude by model.scale=1e-4 according to pyTMD/pull/243, fill NA/n7
 
     -- small fix TypeError: expected unicode string, found 15 problem in constituents, fill NA/n11
+
+#### ver 0.0.8 Add nearest method to locate points by Zarr; Validate api results with Jen's paper